requirements.txt
agate==1.7.1
altair==5.3.0
annotated-types==0.6.0
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
DateTime==5.5
db-dtypes==1.2.0
dbt-bigquery==1.7.7
dbt-core==1.7.12
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.4.4
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.18.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.20.1
google-cloud-core==2.4.1
google-cloud-dataproc==5.9.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.7
importlib-metadata==6.11.0
isodate==0.6.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
leather==0.4.0
Logbook==1.5.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mashumaro==3.12
mdurl==0.1.2
minimal-snowplow-tracker==0.0.2
more-itertools==10.2.0
msgpack==1.0.8
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
pandas-gbq==0.23.0
parsedatetime==2.6
pathspec==0.11.2
pillow==10.3.0
plotly==5.22.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydata-google-auth==1.8.2
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
setuptools==70.0.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sqlparse==0.4.4
streamlit==1.35.0
tenacity==8.3.0
text-unidecode==1.3
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
zipp==3.18.1
zope.interface==6.4.post2
